# backend/api/swagger/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: fess932@gmail.com
  description: 'Сервер ReStar'
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Заказы
  version: 1.0.0
...
servers:
- url: http://localhost:3000/
- url: https://restar26.site/
...
paths:
...
  /pet/{petId}/uploadImage:
    post:
      operationId: uploadFile
      parameters:
      - description: ID of pet to update
        in: path
        name: petId
        required: true
        schema:
          format: int64
          type: integer
          x-exportParamName: PetId
        x-exportParamName: PetId
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                additionalMetadata:
                  description: Additional data to pass to server
                  type: string
                  x-exportParamName: AdditionalMetadata
                  x-optionalDataType: String
                file:
                  description: file to upload
                  format: binary
                  type: string
                  x-exportParamName: File
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: successful operation
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: uploads an image
      tags:
      - pet
...
components:
  schemas:
...
    ApiResponse:
      example:
        code: 0
        type: type
        message: message
      properties:
        code:
          format: int32
          type: integer
        type:
          type: string
        message:
          type: string
      type: object
  securitySchemes:
    petstore_auth:
      flows:
        implicit:
          authorizationUrl: http://petstore.swagger.io/oauth/dialog
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey